Refactor /ledger, fix broken tests, finally sort out deps
