+++ __mocks__/@ledgerhq/hw-transport-node-hid.js
@@ -1,0 +1,1 @@
+module.exports = {}
