+++ __tests__/setupTests.js
@@ -9,6 +9,13 @@ configure({ adapter: new Adapter() })
 
 beforeAll(() => {
   api.setSwitchFreeze(true)
+
+  jest.doMock('../node_modules/usb/usb.js', () => {
+    return {
+      __esModule: true,
+      default: '',
+    }
+  })
 })
 
 afterAll(() => {
