+++ app/actions/authActions.js
@@ -8,7 +8,7 @@ import { bindArgsFromN } from '../util/bindHelpers'
 import { resetBalanceState } from './balancesActions'
 import { upgradeNEP6AddAddresses } from '../core/account'
 import { validatePassphraseLength } from '../core/wallet'
-import { ledgerNanoSCreateSignatureAsync } from '../ledger/ledgerNanoS'
+import { legacySignWithLedger } from '../ledger/neonLedger'
 
 type WifLoginProps = {
   wif: string,
@@ -130,11 +130,7 @@ export const ledgerLoginActions = createActions(
     return {
       publicKey,
       address: walletAccount.address,
-      signingFunction: bindArgsFromN(
-        ledgerNanoSCreateSignatureAsync,
-        3,
-        account,
-      ),
+      signingFunction: bindArgsFromN(legacySignWithLedger, 3, account),
       isHardwareLogin: true,
       hasInternetConnectivity,
     }
